Fix warnings on x86-64 (Tres Melton).
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@15203 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/areas.c
@@ -265,7 +265,7 @@ SetCurrentArea(int ax, int ay)
    /* re-focus on a new ewin on that new desktop area */
    FocusNewDesk();
 
-   ModulesSignal(ESIGNAL_AREA_SWITCH_DONE, (void *)DesksGetCurrent());
+   ModulesSignal(ESIGNAL_AREA_SWITCH_DONE, ((void *)(long)DesksGetCurrent()));
 
    /* update which "edge flip resistance" detector windows are visible */
    EdgeWindowsShow();
